+++ app/src/main/kotlin/kttp/http/protocol/HttpResponse.kt
@@ -30,7 +30,7 @@ class HttpResponse(val statusLine: StatusLine, val headers: HttpHeaders, val bod
         fun fromStatus(httpStatus: HttpStatus, headers: HttpHeaders = HttpHeaders(), body: String = ""): HttpResponse {
             val statusLine = StatusLine(HttpVersion.DEFAULT_VERSION, httpStatus)
             if(!headers.hasContentLength())
-                headers.withContentLength(body.length)
+                headers.withContentLength(body.length.toLong())
             return HttpResponse(statusLine, headers, HttpBody.fromString(body))
         }
